Fix failures of unittest

+++ tests/test_vimrc_creator.rb
@@ -3,11 +3,20 @@ require_relative '../lib/vimrc_creator'
 
 class TestVimrcCreator < Test::Unit::TestCase
   @@filepath = './test_vimrc'
+  @@filepath_fixed = './test_fixed_vimrc'
 
   def teardown
-    if File.exist?(@@filepath)
-      File.unlink(@@filepath)
-    end
+    File.unlink(@@filepath) if File.exist?(@@filepath)
+    File.unlink(@@filepath_fixed) if File.exist?(@@filepath_fixed)
+  end
+
+  # 固定設定の配列を返すメソッド
+  # -> 不正な値を登録しようとした際に
+  #    固定設定のみ反映されているかを確認する際に使う
+  def fixed_options
+    @vimrc_creator = VimFactory::VimrcCreator.new({}, @@filepath_fixed)
+    @vimrc_creator.create
+    File.read(@@filepath_fixed).split("\n")
   end
 
   # toggle_optionの書き込み処理
@@ -15,12 +24,10 @@ class TestVimrcCreator < Test::Unit::TestCase
   def test_create_with_toggle_option_true
     @contents = { 'ruler' => true }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      line = file.gets
-      assert_equal(line, "set ruler\n")
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file.include?("set ruler"), true)
   end
 
   # toggle_optionの書き込み処理
@@ -28,12 +35,10 @@ class TestVimrcCreator < Test::Unit::TestCase
   def test_create_with_toggle_option_false
     @contents = { 'ruler' => false }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      line = file.gets
-      assert_equal(line, "set noruler\n")
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file.include?("set noruler"), true)
   end
 
   # toggle_optionの書き込み処理
@@ -41,64 +46,56 @@ class TestVimrcCreator < Test::Unit::TestCase
   def test_create_with_invalid_toggle_option
     @contents = { 'ruler' => 'foo' }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      line = file.gets
-      assert_equal(line, nil)
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file, fixed_options)
   end
 
   # string_optionの書き込み処理
   # キーが妥当、値が妥当
   def test_create_with_valid_string_option
-    @contents = { 'encoding' => 'sjis' }
+    @contents = { 'syntax' => 'on' }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      line = file.gets
-      assert_equal(line, "set encoding=sjis\n")
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file.include?("set syntax=on"), true)
   end
+
 
   # string_optionの書き込み処理
-  # キーが妥当、値が不正
+  # キーが妥当、値が不正 => 固定設定のみ
   def test_create_with_invalid_string_option
-    @contents = { 'encoding' => 'foo' }
+    @contents = { 'syntax' => 'foo' }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      line = file.gets
-      assert_equal(line, nil)
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file, fixed_options)
   end
+
 
   # string_optionの書き込み処理
   # キーが不正、値が妥当
   def test_create_with_invalid_string_option2
-    @contents = { 'encoding___' => 'sjis' }
+    @contents = { 'syntax__' => 'on' }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      line = file.gets
-      assert_equal(line, nil)
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file, fixed_options)
   end
 
   # string_optionの書き込み処理
   # キーが不正、値が不正
   def test_create_with_invalid_string_option3
-    @contents = { 'encoding___' => 'foo' }
+    @contents = { 'syntax__' => 'foo' }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      line = file.gets
-      assert_equal(line, nil)
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file, fixed_options)
   end
 
   # number_optionの書き込み処理
@@ -106,12 +103,10 @@ class TestVimrcCreator < Test::Unit::TestCase
   def test_create_with_valid_number_option
     @contents = { 'tabstop' => 10 }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      line = file.gets
-      assert_equal(line, "set tabstop=10\n")
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file.include?("set tabstop=10"), true)
   end
 
   # number_optionの書き込み処理
@@ -119,12 +114,10 @@ class TestVimrcCreator < Test::Unit::TestCase
   def test_create_with_min_number_option4
     @contents = { 'tabstop' => 0 }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      line = file.gets
-      assert_equal(line, "set tabstop=0\n")
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file.include?("set tabstop=0"), true)
   end
 
   # number_optionの書き込み処理
@@ -132,12 +125,10 @@ class TestVimrcCreator < Test::Unit::TestCase
   def test_create_with_max_number_option
     @contents = { 'tabstop' => 100 }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      line = file.gets
-      assert_equal(line, "set tabstop=100\n")
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file.include?("set tabstop=100"), true)
   end
 
   # number_optionの書き込み処理
@@ -145,12 +136,10 @@ class TestVimrcCreator < Test::Unit::TestCase
   def test_create_with_invalid_number_option
     @contents = { 'tabstop___' => 100 }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      line = file.gets
-      assert_equal(line, nil)
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file, fixed_options)
   end
 
   # number_optionの書き込み処理
@@ -158,12 +147,10 @@ class TestVimrcCreator < Test::Unit::TestCase
   def test_create_with_invalid_number_option2
     @contents = { 'tabstop' => -1 }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      line = file.gets
-      assert_equal(line, nil)
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file, fixed_options)
   end
 
   # number_optionの書き込み処理
@@ -171,12 +158,10 @@ class TestVimrcCreator < Test::Unit::TestCase
   def test_create_with_invalid_number_option3
     @contents = { 'tabstop' => 101 }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      line = file.gets
-      assert_equal(line, nil)
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file, fixed_options)
   end
 
   # colorschemeの書き込み処理
@@ -184,12 +169,10 @@ class TestVimrcCreator < Test::Unit::TestCase
   def test_create_with_valid_colorscheme
     @contents = { 'colorscheme' => 'desert' }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      line = file.gets
-      assert_equal(line, "colorscheme desert\n")
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file.include?("colorscheme desert"), true)
   end
 
   # colorschemeの書き込み処理
@@ -197,12 +180,10 @@ class TestVimrcCreator < Test::Unit::TestCase
   def test_create_with_invalid_colorscheme
     @contents = { 'colorscheme' => 'foo' }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      line = file.gets
-      assert_equal(line, nil)
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file, fixed_options)
   end
 
   # 全オプションの書き込み処理
@@ -214,13 +195,12 @@ class TestVimrcCreator < Test::Unit::TestCase
       'colorscheme' => 'morning'
     }
     @vimrc_creator = VimFactory::VimrcCreator.new(@contents, @@filepath)
-    @vimrc_creator.create()
+    @vimrc_creator.create
 
-    File.open(@@filepath) do |file|
-      assert_equal(file.gets, "set ruler\n")
-      assert_equal(file.gets, "set syntax=on\n")
-      assert_equal(file.gets, "set tabstop=4\n")
-      assert_equal(file.gets, "colorscheme morning\n")
-    end
+    file = File.read(@@filepath).split("\n")
+    assert_equal(file.include?("set ruler"), true)
+    assert_equal(file.include?("set syntax=on"), true)
+    assert_equal(file.include?("set tabstop=4"), true)
+    assert_equal(file.include?("colorscheme morning"), true)
   end
 end
